requirements.txt
flask
influxdb-client
paho-mqtt
